site_vote_photo/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
daphne==3.0.2
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dj-database-url==1.3.0
dj3-cloudinary-storage==0.0.6
Django==4.1.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-channels-notifications==1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-fsm==2.8.1
django-service-objects==0.7.1
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
easy-thumbnails==2.8.5
filelock==3.10.0
flake8==6.0.0
Flask==2.2.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
handler==1.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.4
mypy==1.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pipenv==2023.2.18
pipfile==0.0.2
platformdirs==3.1.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
route==0.1.4
ruamel.yaml==0.17.21
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
stories==5.1.0
toml==0.10.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
vk==3.0
wcwidth==0.2.6
websocket==0.2.1
Werkzeug==2.2.3
whitenoise==6.4.0
zope.event==4.6
zope.interface==5.5.2
